+++ changelog.md
@@ -9,8 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.2] - 2025-04-29
-
 ### Changed
 
 - Gradle dependency updates
@@ -18,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - BL-1340 - Fix alias for VerifySCryptHash
+- BL-1339 - Add support for ACF stripping scrypt version prefix from hashes
 
 ## [v1.0.1] - 2025-03-01
 
@@ -33,4 +32,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/ortus-boxlang/bx-password-encrypt/compare/v1.0.2...HEAD
 
-[1.0.2]: https://github.com/ortus-boxlang/bx-password-encrypt/compare/f6a4b33af6c35a02578f2ec87b03a33bf12455a0...v1.0.2